agent/requirements.txt
requests
Pillow
pywin32
pyinstaller
